+++ app/src/main/java/com/health/myhealth/utils/ListenerEventSensor.java
@@ -1,5 +1,5 @@
 package com.health.myhealth.utils;
 
 public interface ListenerEventSensor {
-    void eventSensor(int step, int run, int sleep, double calo, double quangDuong);
+    void eventSensor(int step, int run, long sleep, double calo, double quangDuong);
 }
